+++ config/routes/cashflows_router.js
@@ -21,30 +21,6 @@ module.exports = (app) => {
 
   router.route("/")
     .get(CashflowsController.index)
-  // app.post("/cashflows", (req, res) => {
-  //   if (req.body.constructor === Object) {
-  //     let keys = Object.keys(req.body)
-  //     if (keys.includes("amount") && keys.includes("description")) {
-  //       req.app.locals.db.run(`
-  //         INSERT INTO cashflows (
-  //           amount,
-  //           description
-  //         ) VALUES (?, ?);
-  //       `, [req.body.amount, req.body.description], () => {
-  //         res.redirect("/")
-  //       })
-  //     } else {
-  //       res.sendStatus(400)
-  //     }
-  //   } else {
-  //     res.sendStatus(400)
-  //   }
-  // })
-  // app.delete("/cashflows/:id", (req, res) => {
-  //   req.app.locals.db.run(`DELETE FROM cashflows WHERE id=${req.params.id};`, () => {
-  //     res.redirect("/")
-  //   })
-  // })
 
   /**
    * Export router.