+++ db/migrations/20171106201130_create_cashflows.js
@@ -1,7 +1,7 @@
 "use strict"
 
 module.exports = async (db) => {
-  console.log('migrations/20171104201130_create_cashflows.js')
+  console.log('migrations/20171106201130_create_cashflows.js')
 
   try {
     await db.query(`
@@ -9,6 +9,7 @@ BEGIN TRANSACTION;
 
 CREATE TABLE IF NOT EXISTS cashflows (
   id integer PRIMARY KEY,
+  user_id integer NOT NULL,
   amount integer DEFAULT 0,
   description text
 );