+++ db/seed.sql
@@ -1,0 +1,5 @@
+BEGIN TRANSACTION;
+
+INSERT INTO users (id, balance, currency) VALUES (0, 0, 0);
+
+COMMIT;